nem-dashboard-backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
aiofiles>=23.0.0
aiohttp>=3.8.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0